#3 refactor: Modal 구조 변경 및 디렉토리 구조 변경
- Modal을 더 단순한 구조로 변경
- RegionModal을 상태에 따라 달라지는 두 개의 컴포넌트로 분리
- Region과 Category에 속하는 Modal들을 각 폴더로 분리
- 코딩 컨벤션 피드백에 맞게 수정
