--- src/components/common/Modal/index.ts
@@ -1,2 +1,0 @@
-export { default as RegionModal } from "./RegionModal";
-export { default as CategoryModal } from "./CategoryModal";
